Using cEditorUndoRedoHistory
