+++ src/gamestates/cEditorState.cpp
@@ -1,4 +1,5 @@
 #include "gamestates/cEditorState.h"
+#include "gamestates/cEditorState/cEditorUndoRedoHistory.h"
 #include "controls/eKeyAction.h"
 #include "gui/GuiBar.h"
 #include "gui/GuiStateButton.h"
@@ -20,6 +21,7 @@
 #include <fstream>
 #include <algorithm>
 #include <cctype>
+#include <memory>
 
 const int heightBarSize = 48;
 const int heightButtonSize = 40;
@@ -46,6 +48,8 @@ cEditorState::cEditorState(sGameServices* services)
     assert(m_settings != nullptr);
     assert(m_interface != nullptr);
     assert(m_textDrawer != nullptr);
+
+    m_undoRedo = std::make_unique<cEditorUndoRedoHistory>();
 
     const cRectangle &selectRect = cRectangle(0, 0, m_settings->getScreenW(), heightBarSize);
     const cRectangle &modifRect = cRectangle(m_settings->getScreenW()-heightBarSize, heightBarSize, heightBarSize, m_settings->getScreenH()-heightBarSize);
@@ -74,89 +78,7 @@ cEditorState::cEditorState(sGameServices* services)
 
 cEditorState::~cEditorState()
 {
-
 }
-
-// ---- Undo/Redo helpers ---------------------------------------------------
-
-void cEditorState::beginRecordGroup()
-{
-    m_undoStack.push_back(Marker{});
-    m_redoStack.clear();
-}
-
-void cEditorState::recordTileChange(int col, int row, int oldTileID, int newTileID)
-{
-    m_undoStack.push_back(TileChange{col, row, oldTileID, newTileID});
-}
-
-void cEditorState::recordStartCellChange(int playerID, cPoint oldPos, cPoint newPos)
-{
-    m_undoStack.push_back(StartCellChange{playerID, oldPos, newPos});
-}
-
-void cEditorState::applyUndo()
-{
-    if (m_undoStack.empty()) return;
-
-    // Collect entries back to the most recent Marker
-    std::vector<EditorChange> group;
-    while (!m_undoStack.empty() && !std::holds_alternative<Marker>(m_undoStack.back())) {
-        group.push_back(m_undoStack.back());
-        m_undoStack.pop_back();
-    }
-    if (!m_undoStack.empty()) {
-        m_undoStack.pop_back(); // pop the Marker itself
-    }
-
-    if (group.empty()) return;
-
-    // Push redo marker + reversed changes
-    m_redoStack.push_back(Marker{});
-    for (auto &change : group) {
-        if (auto *tc = std::get_if<TileChange>(&change)) {
-            (*m_mapData)[tc->row][tc->col] = tc->oldTileID;
-            m_redoStack.push_back(TileChange{tc->col, tc->row, tc->oldTileID, tc->newTileID});
-        } else if (auto *sc = std::get_if<StartCellChange>(&change)) {
-            startCells[sc->playerID] = sc->oldPos;
-            m_redoStack.push_back(StartCellChange{sc->playerID, sc->oldPos, sc->newPos});
-        }
-    }
-
-    m_hasChanged = true;
-}
-
-void cEditorState::applyRedo()
-{
-    if (m_redoStack.empty()) return;
-
-    std::vector<EditorChange> group;
-    while (!m_redoStack.empty() && !std::holds_alternative<Marker>(m_redoStack.back())) {
-        group.push_back(m_redoStack.back());
-        m_redoStack.pop_back();
-    }
-    if (!m_redoStack.empty()) {
-        m_redoStack.pop_back(); // pop the Marker
-    }
-
-    if (group.empty()) return;
-
-    // Push undo marker + re-applied changes
-    m_undoStack.push_back(Marker{});
-    for (auto &change : group) {
-        if (auto *tc = std::get_if<TileChange>(&change)) {
-            (*m_mapData)[tc->row][tc->col] = tc->newTileID;
-            m_undoStack.push_back(TileChange{tc->col, tc->row, tc->oldTileID, tc->newTileID});
-        } else if (auto *sc = std::get_if<StartCellChange>(&change)) {
-            startCells[sc->playerID] = sc->newPos;
-            m_undoStack.push_back(StartCellChange{sc->playerID, sc->oldPos, sc->newPos});
-        }
-    }
-
-    m_hasChanged = true;
-}
-
-// --------------------------------------------------------------------------
 
 void cEditorState::populateSelectBar()
 {
@@ -536,10 +458,14 @@ void cEditorState::onNotifyKeyboardEvent(const cKeyboardEvent &event)
             m_displayAxes = !m_displayAxes;
         }
         if (event.isAction(eKeyAction::EDITOR_UNDO)) {
-            applyUndo();
+            if (m_undoRedo->applyUndo(*m_mapData, startCells)) {
+                m_hasChanged = true;
+            }
         }
         if (event.isAction(eKeyAction::EDITOR_REDO)) {
-            applyRedo();
+            if (m_undoRedo->applyRedo(*m_mapData, startCells)) {
+                m_hasChanged = true;
+            }
         }
     }
 
@@ -596,8 +522,7 @@ void cEditorState::loadMap(s_PreviewMap* map)
     m_displayGrid = false;
     m_displayAxes = false;
     m_hasChanged = false;
-    m_undoStack.clear();
-    m_redoStack.clear();
+    m_undoRedo->clear();
 }
 
 void cEditorState::setCursorSize(int value)
@@ -925,11 +850,11 @@ void cEditorState::modifyTile(int posX, int posY, int tileID)
             }
 
             if (!hasRecordedGroup) {
-                beginRecordGroup();
+                m_undoRedo->beginRecordGroup();
                 hasRecordedGroup = true;
             }
 
-            recordTileChange(brushTileX, brushTileY, oldTileID, tileID);
+            m_undoRedo->recordTileChange(brushTileX, brushTileY, oldTileID, tileID);
             (*m_mapData)[brushTileY][brushTileX] = tileID;
             m_hasChanged = true;
         }
@@ -947,8 +872,8 @@ void cEditorState::modifyStartCell(int posX, int posY, int startCellID)
         cPoint oldPos = startCells[startCellID];
         cPoint newPos = {tileX, tileY};
         if (oldPos.x == newPos.x && oldPos.y == newPos.y) return;
-        beginRecordGroup();
-        recordStartCellChange(startCellID, oldPos, newPos);
+        m_undoRedo->beginRecordGroup();
+        m_undoRedo->recordStartCellChange(startCellID, oldPos, newPos);
         startCells[startCellID] = newPos;
         m_hasChanged = true;
     }
@@ -957,7 +882,7 @@ void cEditorState::modifyStartCell(int posX, int posY, int startCellID)
 void cEditorState::modifySymmetricArea(Direction dir)
 {
     // std::cout << "modifySymmetricArea " << static_cast<int>(dir) << std::endl;
-    beginRecordGroup();
+    m_undoRedo->beginRecordGroup();
     switch (dir) {
         case Direction::bottom:
             for (size_t j = 1; j < (m_mapData->getRows())/2; j++) {
@@ -966,7 +891,7 @@ void cEditorState::modifySymmetricArea(Direction dir)
                     int oldVal = (*m_mapData)[mirrorJ][i];
                     int newVal = (*m_mapData)[j][i];
                     if (oldVal != newVal) {
-                        recordTileChange(i, mirrorJ, oldVal, newVal);
+                        m_undoRedo->recordTileChange(i, mirrorJ, oldVal, newVal);
                         (*m_mapData)[mirrorJ][i] = newVal;
                     }
                 }
@@ -979,7 +904,7 @@ void cEditorState::modifySymmetricArea(Direction dir)
                     int oldVal = (*m_mapData)[j][i];
                     int newVal = (*m_mapData)[mirrorJ][i];
                     if (oldVal != newVal) {
-                        recordTileChange(i, j, oldVal, newVal);
+                        m_undoRedo->recordTileChange(i, j, oldVal, newVal);
                         (*m_mapData)[j][i] = newVal;
                     }
                 }
@@ -992,7 +917,7 @@ void cEditorState::modifySymmetricArea(Direction dir)
                     int oldVal = (*m_mapData)[j][mirrorI];
                     int newVal = (*m_mapData)[j][i];
                     if (oldVal != newVal) {
-                        recordTileChange(mirrorI, j, oldVal, newVal);
+                        m_undoRedo->recordTileChange(mirrorI, j, oldVal, newVal);
                         (*m_mapData)[j][mirrorI] = newVal;
                     }
                 }
@@ -1005,7 +930,7 @@ void cEditorState::modifySymmetricArea(Direction dir)
                     int oldVal = (*m_mapData)[j][i];
                     int newVal = (*m_mapData)[j][mirrorI];
                     if (oldVal != newVal) {
-                        recordTileChange(i, j, oldVal, newVal);
+                        m_undoRedo->recordTileChange(i, j, oldVal, newVal);
                         (*m_mapData)[j][i] = newVal;
                     }
                 }
