+++ src/gamestates/cEditorState.h
@@ -9,8 +9,6 @@
 #include <memory>
 #include <array>
 #include <string>
-#include <variant>
-#include <vector>
 
 class Texture;
 class Graphics;
@@ -20,6 +18,7 @@ class GuiBar;
 class GuiButtonGroup;
 class cTextDrawer;
 class cGameInterface;
+class cEditorUndoRedoHistory;
 
 class cEditorState : public cGameState {
 public:
@@ -36,21 +35,6 @@ public:
 
     eGameStateType getType() override;
 private:
-    // ---- Undo/Redo -------------------------------------------------------
-    struct Marker {};
-    struct TileChange    { int col, row, oldTileID, newTileID; };
-    struct StartCellChange { int playerID; cPoint oldPos, newPos; };
-    using EditorChange = std::variant<Marker, TileChange, StartCellChange>;
-
-    std::vector<EditorChange> m_undoStack;
-    std::vector<EditorChange> m_redoStack;
-
-    void recordTileChange(int col, int row, int oldTileID, int newTileID);
-    void recordStartCellChange(int playerID, cPoint oldPos, cPoint newPos);
-    void beginRecordGroup();
-    void applyUndo();
-    void applyRedo();
-    // -----------------------------------------------------------------------
     Graphics *m_gfxdata = nullptr;
     Graphics *m_gfxeditor = nullptr;
     cGameSettings *m_settings = nullptr;
@@ -101,6 +85,8 @@ private:
     std::unique_ptr<GuiButtonGroup> m_topologyGroup;
     std::unique_ptr<GuiButtonGroup> m_startCellGroup;
     std::unique_ptr<GuiButtonGroup> m_symmetricGroup;
+
+    std::unique_ptr<cEditorUndoRedoHistory> m_undoRedo;
 
     cRectangle mapSizeArea;
     int tileLenSize = 16;
